# src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Kaburlu Backend API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:4000"
    },
    {
      "url": "https://influapi.kaburlumedia.com"
    }
  ],
...
  "paths": {
...
    "/api/profile-builder/pack/{id}": {
      "get": {
        "tags": [
          "Profile Builder"
        ],
        "summary": "Get profile pack",
        "security": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfilePack"
                }
              }
            }
          },
          "404": { "description": "Not found" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "ProfilePack": {
        "type": "object",
        "properties": {
          "packId": { "type": "string" },
          "name": { "type": "string" },
          "niche": { "type": "string" },
          "city": { "type": "string" },
          "shortBio": { "type": "string" },
          "longBio": { "type": "string" },
          "jsonLd": { "type": "object" },
          "images": { "type": "object", "properties": { "avatar": { "type": "string", "nullable": true }, "gallery": { "type": "array", "items": { "type": "string" } } } },
          "ai": { "type": "boolean" }
        }
      },
...
    }
  }
}